📑 Add proof, solution, exercise defaults to typst export
