+++ .changeset/curvy-geese-retire.md
@@ -1,0 +1,5 @@
+---
+"myst-to-typst": patch
+---
+
+Add exercise and solution handlers
